+++ src/main/java/com/getindata/connectors/http/internal/HttpLoggingLevelType.java
@@ -1,0 +1,15 @@
+package com.getindata.connectors.http.internal;
+
+public enum HttpLoggingLevelType {
+    MIN,
+    REQRESPONSE,
+    MAX;
+
+    public static HttpLoggingLevelType valueOfStr(String code) {
+        if (code == null) {
+            return MIN;
+        } else {
+            return valueOf(code);
+        }
+    }
+}