+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClient.java
@@ -31,6 +31,7 @@ import org.apache.flink.util.StringUtils;
 import com.getindata.connectors.http.HttpPostRequestCallback;
 import com.getindata.connectors.http.HttpStatusCodeValidationFailedException;
 import com.getindata.connectors.http.internal.HeaderPreprocessor;
+import com.getindata.connectors.http.internal.HttpLogger;
 import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.retry.HttpClientWithRetry;
 import com.getindata.connectors.http.internal.retry.RetryConfigProvider;
@@ -63,6 +64,7 @@ public class JavaNetHttpPollingClient implements PollingClient {
     private final HttpLookupConfig options;
     private final Set<Integer> ignoredErrorCodes;
     private final boolean continueOnError;
+    private final HttpLogger httpLogger;
 
     public JavaNetHttpPollingClient(
             HttpClient httpClient,
@@ -89,6 +91,7 @@ public class JavaNetHttpPollingClient implements PollingClient {
                 .retryConfig(RetryConfigProvider.create(config))
                 .responseChecker(new HttpResponseChecker(successCodes, errorCodes))
                 .build();
+        this.httpLogger = HttpLogger.getHttpLogger(options.getProperties());
     }
 
     public void open(FunctionContext context) {
@@ -118,15 +121,20 @@ public class JavaNetHttpPollingClient implements PollingClient {
         HttpResponse<String> response =null;
         HttpRowDataWrapper httpRowDataWrapper = null;
         try {
+            httpLogger.logRequest(request.getHttpRequest());
             response = httpClient.send(
                 () -> updateHttpRequestIfRequired(request, oidcProcessor), BodyHandlers.ofString());
+            httpLogger.logResponse(response);
         } catch (HttpStatusCodeValidationFailedException e) {
+            // log if we fail for status code reasons.
+            httpLogger.logResponse((HttpResponse<String>) e.getResponse());
             // Case 1 http non successful response
             if (!this.continueOnError) throw e;
             // use the response in the Exception
             response = (HttpResponse<String>) e.getResponse();
             httpRowDataWrapper = processHttpResponse(response, request, true);
         } catch (Exception e) {
+            httpLogger.logExceptionResponse(request, e);
             // Case 2 Exception occurred
             if (!this.continueOnError) throw e;
             String errMessage =  e.getMessage();
@@ -210,9 +218,9 @@ public class JavaNetHttpPollingClient implements PollingClient {
             boolean isError) throws IOException {
 
         this.httpPostRequestCallback.call(response, request, "endpoint", Collections.emptyMap());
+        this.httpLogger.logResponse(response);
         var responseBody = response.body();
 
-        log.debug("Received status code [{}] for RestTableSource request", response.statusCode());
         if (!isError && (StringUtils.isNullOrWhitespaceOnly(responseBody) || ignoreResponse(response))) {
             return HttpRowDataWrapper.builder()
                     .data(Collections.emptyList())